Add simple command line parser for TestLauncher
The CommandLineParser is able to parse command line options in the form of
`--option1 param1 param2 --option2 --option3 param3`.
Each option starts with two dashes and any arguments not beginning with `--`
are treated as parameters for the last option that was encountered. So `param1`
and `param2` are parameters for `option1``,  `option2` has no parameters and
`option3` has 1 parameter `param3`.

Currently there is no schema validation or type conversion, so checking that
is left up to the caller. To ease that, `CommandLineParser` provides some
utility functions:
  - `isPresent(String name)` checks if a given option was present on the
    command line
  - `getMissingOptions(String ... names)` returns an array of arguments which
    were not present but should have been according to `names`
  - checkRequiredOptions(String ... names) checks if all options named were
    given and return `false` if any were missing while also printing an error
    message to `System.err` for each missing option.

Furthermore `getParametersFor(String name)` can be used to obtain an Array of
an option's associated parameters and `getFirstParameterFor(String name)`
provides a convenient way to access an option's first parameter. This is
especially useful for single parameter options.

Using this parser, `TstLauncher` has been rewritten to accept the target deep
file and target configuration as command line parameters instead of
uncommenting the corresponding line in `TestLauncher` itself. The new command
line options are:
  - `--deepfile`: the deep configuration file to use `555ExampleProject.deep`
  - `--targetconfig`: the target configuration to use, e.g. `BootFromRam`